Consolidate mining crate systems after review
Merge points/timer unlock into MiningCrateSystem and force-power-off into
security. Collapse visual/wire enums into one file. Keep optional components.
