--- Content.Goobstation.Shared/MiningCrate/MiningCrateSecurityWireKeys.cs
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-using Robust.Shared.Serialization;
-
-namespace Content.Goobstation.Shared.MiningCrate;
-
-[Serializable, NetSerializable]
-public enum MiningCrateSecurityWireStatus : byte
-{
-    SecurityIndicator,
-    LockIndicator,
-    SirenIndicator,
-    BoomIndicator,
-}
